requirements.txt
asgiref==3.2.7
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dialogflow==1.0.0
Django==3.0.3
django-google-maps==0.12.1
django-jquery==3.1.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
geographiclib==1.50
geopy==1.22.0
google-api-core==1.20.0
google-auth==1.17.2
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
phonenumbers==8.12.5
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
twilio==6.42.0
urllib3==1.25.9
whitenoise==5.1.0
